+++ app/controllers/tasks/leads/dichotomous_key_controller.rb
@@ -1,0 +1,4 @@
+class Tasks::Leads::DichotomousKeyController < ApplicationController
+  include TaskControllerConfiguration
+
+end