+++ app/javascript/vue/routes/routes.js
@@ -31,6 +31,7 @@ const RouteNames = {
   BrowseOtu: '/tasks/otus/browse',
   BrowseSound: '/tasks/sounds/browse',
   ContentEditor: '/tasks/content/editor',
+  DichotomousKey: '/tasks/leads/dichotomous_key',
   DigitizeTask: '/tasks/accessions/comprehensive',
   DwcDashboard: '/tasks/dwc/dashboard',
   DwcImport: '/tasks/dwca_import/index',
@@ -92,6 +93,7 @@ const RouteNames = {
   PrintLabel: '/tasks/labels/print_labels',
   ProjectVocabulary: '/tasks/metadata/vocabulary/project_vocabulary',
   ShowLead: '/tasks/leads/show',
+  SourceHub: '/tasks/sources/hub',
   TypeMaterial: '/tasks/type_material/edit_type_material',
   UnifyObjects: '/tasks/unify/objects',
   UnifyPeople: '/tasks/unify/people'