+++ app/javascript/vue/tasks/news/browse/components/Viewer/NewsViewer.vue
@@ -23,9 +23,7 @@
       v-text="news.title"
     />
     <div>
-      <div class="news-meta">
-        {{ news.creator }} — {{ formatDate(utcToLocal(news.createdAt)) }}
-      </div>
+      <div class="news-meta">{{ news.creator }} — {{ news.createdAt }}</div>
     </div>
     <div v-html="news.body" />
     <NewsViewerBack
@@ -37,7 +35,7 @@
 
 <script setup>
 import { computed } from 'vue'
-import { utcToLocal, formatDate, isCurrentUserAdministrator } from '@/helpers'
+import { isCurrentUserAdministrator } from '@/helpers'
 import newsColors from '../../constants/newsColors'
 import NewsViewerBack from './NewsViewerBack.vue'
 import VBadge from '@/components/ui/VBadge/VBadge.vue'