+++ app/javascript/vue/tasks/sources/new_source/components/bibtex/copyright.vue
@@ -1,27 +1,17 @@
 <template>
-  <div class="field">
-    <label>Copyright</label><br>
+  <div class="label-above">
+    <label>Copyright</label>
     <input
       type="text"
-      v-model="source.copyright"/>
+      v-model="source.copyright"
+      @change="() => (source.isUnsaved = true)"
+    />
   </div>
 </template>
 
-<script>
-
-import { GetterNames } from '../../store/getters/getters'
-import { MutationNames } from '../../store/mutations/mutations'
-
-export default {
-  computed: {
-    source: {
-      get () {
-        return this.$store.getters[GetterNames.GetSource]
-      },
-      set (value) {
-        this.$store.commit(MutationNames.SetSource, value)
-      }
-    }
-  }
-}
+<script setup>
+const source = defineModel({
+  type: Object,
+  required: true
+})
 </script>