+++ spec/lib/queries/taxon_name/filter_spec.rb
@@ -657,6 +657,22 @@ describe Queries::TaxonName::Filter, type: :model, group: [:nomenclature] do
     expect(query.all.map(&:id)).to contain_exactly()
   end
 
+  context 'availability' do
+    before(:each) {
+      TaxonNameClassification::Iczn::Unavailable.create!(taxon_name: species)
+    }
+
+    specify '#availability 1' do
+      query.availability = true
+      expect(query.all.map(&:id).size).to eq(3)
+    end
+
+    specify '#validity 2' do
+      query.availability = false
+      expect(query.all.map(&:id)).to contain_exactly(species.id)
+    end
+  end
+
   specify 'all filters combined' do
     Citation.create!(citation_object: species, source: FactoryBot.create(:valid_source))
     Otu.create!(taxon_name: species)
@@ -684,5 +700,26 @@ describe Queries::TaxonName::Filter, type: :model, group: [:nomenclature] do
     query.taxon_name_type = 'Protonym'
 
     expect(query.all.map(&:id)).to contain_exactly(species.id)
+  end
+
+  specify 'otu_query with venn subquery subtraction' do
+    o1 = Otu.create!(name: 'Otu One', taxon_name: species)
+    o2 = Otu.create!(name: 'Otu Two', taxon_name: genus)
+
+    venn_query_params = {
+      'otu_id' => [o1.id]
+    }
+    venn_url = "http://localhost:3000/otus/filter.json?#{venn_query_params.to_query}"
+
+    q = Queries::TaxonName::Filter.new(
+      otu_query: {
+        otu_id: [o1.id, o2.id],
+        venn: venn_url,
+        venn_mode: 'a',  # A - B = [o1, o2] - [o1] = [o2]
+        venn_ignore_pagination: true
+      }
+    )
+
+    expect(q.all.map(&:id)).to contain_exactly(genus.id)
   end
 end